+++ src/views/AdminLogin.vue
@@ -1,120 +1,12 @@
-<template>
-  <div class="fill-height" id='mainContainer'>
-      <v-container class="fill-height">
-        <v-row align="center" justify="center">
-            <v-card class="d-flex align-center" color="white">
-              <v-row class="mx-10 my-10">
-                <v-col cols="12" md="12">
-                  <h3 class="my-5">Iniciar sesión de Administrador</h3>
-                  <v-text-field
-                    v-model="userLogin.email"
-                    :rules="[rules.required]"
-                    label="Correo electrónico"
-                    single-line
-                    solo
-                  ></v-text-field>
-                  <v-text-field 
-                    v-model="userLogin.password"
-                    :append-icon="show1 ? 'visibility' : 'visibility_off'"
-                    :rules="[rules.required]"
-                    :type="show1 ? 'text' : 'password'"
-                    name="passLogin"
-                    label="Contraseña "                    
-                    solo
-                    @click:append="show1 = !show1"
-                  ></v-text-field>
-                  <v-btn large block color="primary" @click="logIn(userLogin.email,userLogin.password)">Entrar</v-btn>
-                  <v-row>
-                  <v-col cols="12">
-                    <iframe width="100%" height="100%" frameBorder="0" src="https://drive.google.com/file/d/1RaK1u8n79yWg1RD-rLu2BIcL8vqdAkIu/preview" allowfullscreen></iframe>
-                  </v-col>
-                </v-row>                
-                </v-col>
-              </v-row>
-            </v-card>
-        </v-row>
-      </v-container>
-    
-  </div>
+ <template>
+    <Login userType="Administrador" redirectRoute="/admin/home"></Login>
 </template>
 
 <script>
-function test(){
-  window.console.log("TEST")
-}
-
-const helper = require("../helper.js");
-
-import axios from "axios";
-
-  export default {
-    data () {
-      return {
-        isError: false,
-        userLogin: {email:"",password:""},
-        show1: false,
-        rules: {
-          required: value => !!value || 'Required.',
-        },
-      }
-    },
-    methods:{
-
-      logIn(correo,pass){   
-          var temp = {"email":correo.toLowerCase(),"password":pass}
-          const URL = helper.baseURL + "/login";
-          axios
-          .post(URL, temp)
-          .then(response => {
-            localStorage.token = response.data.token
-            this.redirect()
-          }).catch(error=>{  
-            const mensaje = error.response.data.error
-            const errorType = error.response.data.type
-            var iconMessage = "error"
-            var errorTitle = "Error"
-            this.$swal(errorTitle,mensaje, iconMessage)
-          })
-      },
-      redirect() {
-        var route = "/admin/home/";
-        window.open(route, "_self");
-      }
-    }
+import Login from "../components/Login.vue";
+export default {
+  components: {
+    Login
   }
+};
 </script>
-
-
-
-<!-- Add "scoped" attribute to limit CSS to this component only -->
-<style>
-.customSwalBtn{
-		background-color: rgba(214,130,47,1.00);
-    border-left-color: rgba(214,130,47,1.00);
-    border-right-color: rgba(214,130,47,1.00);
-    border: 0;
-    border-radius: 3px;
-    box-shadow: none;
-    color: #fff;
-    cursor: pointer;
-    font-size: 17px;
-    font-weight: 500;
-    margin: 30px 5px 0px 5px;
-    padding: 10px 32px;
-	}
-  h2 {
-    color: black;
-  }
-  #loginSection{
-    border-style: solid;
-    border-width: 0px 1px 0px 0px;
-    border-color: gray;
-  }
-  #mainContainer {
-    /* background by SVGBackgrounds.com */
-    background-color: #7cb342;
-    background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='100%25'%3E%3Cdefs%3E%3ClinearGradient id='a' gradientUnits='userSpaceOnUse' x1='0' x2='0' y1='0' y2='100%25' gradientTransform='rotate(240)'%3E%3Cstop offset='0' stop-color='%237cb342'/%3E%3Cstop offset='1' stop-color='%2342a5f5'/%3E%3C/linearGradient%3E%3Cpattern patternUnits='userSpaceOnUse' id='b' width='540' height='450' x='0' y='0' viewBox='0 0 1080 900'%3E%3Cg fill-opacity='0.1'%3E%3Cpolygon fill='%23444' points='90 150 0 300 180 300'/%3E%3Cpolygon points='90 150 180 0 0 0'/%3E%3Cpolygon fill='%23AAA' points='270 150 360 0 180 0'/%3E%3Cpolygon fill='%23DDD' points='450 150 360 300 540 300'/%3E%3Cpolygon fill='%23999' points='450 150 540 0 360 0'/%3E%3Cpolygon points='630 150 540 300 720 300'/%3E%3Cpolygon fill='%23DDD' points='630 150 720 0 540 0'/%3E%3Cpolygon fill='%23444' points='810 150 720 300 900 300'/%3E%3Cpolygon fill='%23FFF' points='810 150 900 0 720 0'/%3E%3Cpolygon fill='%23DDD' points='990 150 900 300 1080 300'/%3E%3Cpolygon fill='%23444' points='990 150 1080 0 900 0'/%3E%3Cpolygon fill='%23DDD' points='90 450 0 600 180 600'/%3E%3Cpolygon points='90 450 180 300 0 300'/%3E%3Cpolygon fill='%23666' points='270 450 180 600 360 600'/%3E%3Cpolygon fill='%23AAA' points='270 450 360 300 180 300'/%3E%3Cpolygon fill='%23DDD' points='450 450 360 600 540 600'/%3E%3Cpolygon fill='%23999' points='450 450 540 300 360 300'/%3E%3Cpolygon fill='%23999' points='630 450 540 600 720 600'/%3E%3Cpolygon fill='%23FFF' points='630 450 720 300 540 300'/%3E%3Cpolygon points='810 450 720 600 900 600'/%3E%3Cpolygon fill='%23DDD' points='810 450 900 300 720 300'/%3E%3Cpolygon fill='%23AAA' points='990 450 900 600 1080 600'/%3E%3Cpolygon fill='%23444' points='990 450 1080 300 900 300'/%3E%3Cpolygon fill='%23222' points='90 750 0 900 180 900'/%3E%3Cpolygon points='270 750 180 900 360 900'/%3E%3Cpolygon fill='%23DDD' points='270 750 360 600 180 600'/%3E%3Cpolygon points='450 750 540 600 360 600'/%3E%3Cpolygon points='630 750 540 900 720 900'/%3E%3Cpolygon fill='%23444' points='630 750 720 600 540 600'/%3E%3Cpolygon fill='%23AAA' points='810 750 720 900 900 900'/%3E%3Cpolygon fill='%23666' points='810 750 900 600 720 600'/%3E%3Cpolygon fill='%23999' points='990 750 900 900 1080 900'/%3E%3Cpolygon fill='%23999' points='180 0 90 150 270 150'/%3E%3Cpolygon fill='%23444' points='360 0 270 150 450 150'/%3E%3Cpolygon fill='%23FFF' points='540 0 450 150 630 150'/%3E%3Cpolygon points='900 0 810 150 990 150'/%3E%3Cpolygon fill='%23222' points='0 300 -90 450 90 450'/%3E%3Cpolygon fill='%23FFF' points='0 300 90 150 -90 150'/%3E%3Cpolygon fill='%23FFF' points='180 300 90 450 270 450'/%3E%3Cpolygon fill='%23666' points='180 300 270 150 90 150'/%3E%3Cpolygon fill='%23222' points='360 300 270 450 450 450'/%3E%3Cpolygon fill='%23FFF' points='360 300 450 150 270 150'/%3E%3Cpolygon fill='%23444' points='540 300 450 450 630 450'/%3E%3Cpolygon fill='%23222' points='540 300 630 150 450 150'/%3E%3Cpolygon fill='%23AAA' points='720 300 630 450 810 450'/%3E%3Cpolygon fill='%23666' points='720 300 810 150 630 150'/%3E%3Cpolygon fill='%23FFF' points='900 300 810 450 990 450'/%3E%3Cpolygon fill='%23999' points='900 300 990 150 810 150'/%3E%3Cpolygon points='0 600 -90 750 90 750'/%3E%3Cpolygon fill='%23666' points='0 600 90 450 -90 450'/%3E%3Cpolygon fill='%23AAA' points='180 600 90 750 270 750'/%3E%3Cpolygon fill='%23444' points='180 600 270 450 90 450'/%3E%3Cpolygon fill='%23444' points='360 600 270 750 450 750'/%3E%3Cpolygon fill='%23999' points='360 600 450 450 270 450'/%3E%3Cpolygon fill='%23666' points='540 600 630 450 450 450'/%3E%3Cpolygon fill='%23222' points='720 600 630 750 810 750'/%3E%3Cpolygon fill='%23FFF' points='900 600 810 750 990 750'/%3E%3Cpolygon fill='%23222' points='900 600 990 450 810 450'/%3E%3Cpolygon fill='%23DDD' points='0 900 90 750 -90 750'/%3E%3Cpolygon fill='%23444' points='180 900 270 750 90 750'/%3E%3Cpolygon fill='%23FFF' points='360 900 450 750 270 750'/%3E%3Cpolygon fill='%23AAA' points='540 900 630 750 450 750'/%3E%3Cpolygon fill='%23FFF' points='720 900 810 750 630 750'/%3E%3Cpolygon fill='%23222' points='900 900 990 750 810 750'/%3E%3Cpolygon fill='%23222' points='1080 300 990 450 1170 450'/%3E%3Cpolygon fill='%23FFF' points='1080 300 1170 150 990 150'/%3E%3Cpolygon points='1080 600 990 750 1170 750'/%3E%3Cpolygon fill='%23666' points='1080 600 1170 450 990 450'/%3E%3Cpolygon fill='%23DDD' points='1080 900 1170 750 990 750'/%3E%3C/g%3E%3C/pattern%3E%3C/defs%3E%3Crect x='0' y='0' fill='url(%23a)' width='100%25' height='100%25'/%3E%3Crect x='0' y='0' fill='url(%23b)' width='100%25' height='100%25'/%3E%3C/svg%3E");
-    background-attachment: fixed;
-    background-size: cover;
-  }
-</style>
